+++ arch/arm/src/armv8-m/arm_cache.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * arch/arm/src/armv8-m/up_cache.c
+ * arch/arm/src/armv8-m/arm_cache.c
  *
  *   Copyright (C) 2015, 2018-2019 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
