+++ arch/arm/src/armv8-m/arm_schedulesigaction.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * arch/arm/src/armv8-m/up_schedulesigaction.c
+ * arch/arm/src/armv8-m/arm_schedulesigaction.c
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
