+++ arch/arm/src/armv8-m/arm_systemreset.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * arch/arm/src/armv8-m/up_systemreset.c
+ * arch/arm/src/armv8-m/arm_systemreset.c
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
